refactor(StashClient): introducing an interface to provide an alternate implementation for BitBucket cloud
